requirements.txt
streamlit
langchain-openai
langchain-groq
langchain-google-genai
langchain-core
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
python-dotenv
tavily-python